Even though short is the size of what a port is allowed to use, java only supports signed short, which means that pors autogenerated by docker does not work

+++ examples/examples-console/src/main/java/examples/console/example1/App.java
@@ -14,7 +14,7 @@ public class App {
     public static void main(String[] args) throws Exception {
         // Adapter configuration
         final AdapterConfig config = new AdapterConfig("admin", "passwd",
-                "localhost", (short) 6929, 6, defaultSerializer(),
+                "localhost", 6929, 6, defaultSerializer(),
                 new VanillaDataStorageFactory());
 
         // Connect to the server
